requirements.txt
aiohttp==3.7.4.post0
anyio==3.1.0
appdirs==1.4.4
argon2-cffi==20.1.0
astroid==2.5.6
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
Automat==20.2.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.3.0
btrfsutil==5.12.1
CacheControl==0.12.6
ceph==1.0.0
ceph-volume==1.0.0
cephfs==2.0.0
cephfs-shell==0.0.1
cffi==1.14.6
chardet==4.0.0
click==8.0.1
colorama==0.4.4
configobj==5.1.0.dev0
constantly==15.1.0
contextlib2==0.6.0.post1
cryptography==3.4.7
decorator==5.0.9
defusedxml==0.7.1
discord==1.7.3
discord.py==1.7.3
distlib==0.3.2
distro==1.5.0
dnspython==1.16.0
entrypoints==0.3
html5lib==1.1
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
isort==5.8.0
jedi==0.18.0
Jinja2==3.0.1
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-server==1.8.0
jupyterlab==3.0.16
jupyterlab-pygments==0.1.2
jupyterlab-server==2.6.0
lazy-object-proxy==1.6.0
libmsym==0.2.4
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.7.0
msgpack==1.0.2
multidict==5.1.0
nbclassic==0.3.1
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
netsnmp-python==1.0a1
notebook==6.4.0
ordered-set==4.0.2
packaging==20.9
pandocfilters==1.4.3
parso==0.8.2
pep517==0.10.0
pexpect==4.8.0
pickleshare==0.7.5
ply==3.11
powerline-gitstatus==1.3.1
powerline-status==2.8.2
progress==1.5
prometheus-client==0.11.0
prompt-toolkit==3.0.18
psutil==5.8.0
ptyprocess==0.7.0
pycairo==1.20.0
pycparser==2.20
Pygments==2.9.0
PyGObject==3.40.1
pylint==2.8.3
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-sip==12.9.0
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.18.0
pytz==2021.1
pyzmq==22.1.0
rados==2.0.0
rbd==2.0.0
redis==3.5.3
requests==2.25.1
resolvelib==0.5.5
retrying==1.3.3
rgw==2.0.0
Send2Trash==1.5.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.2.1
team==1.0
terminado==0.10.0
terminator==2.1.1
testpath==0.5.0
toml==0.10.2
tornado==6.1
traitlets==5.0.5
Twisted==21.2.0
typing-extensions==3.10.0.0
urllib3==1.26.6
validate==5.1.0.dev0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.0.1
wrapt==1.12.1
yarl==1.6.3
zope.interface==5.4.0
